# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "有个商城api文档",
    "version": "1.0",
    "title": "有个商城API文档",
    "termsOfService": "#",
    "contact": {
      "name": "kyralo",
      "url": "https://amall.kyralo.online",
      "email": "kyralo721@gmail.com"
    },
    "license": {
      "url": "#",
      "name": ""
    }
  },
  "host": "localhost:8888",
  "basePath": "/",
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/account_platforms": {
      "get": {
        "tags": [
          "第三方用户信息"
        ],
        "summary": "分页查询第三方用户信息",
        "operationId": "findByPageUsingGET",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "description": "页号",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页大小",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32",
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbAccountPlatformVO",
              "$ref": "#/definitions/TbAccountPlatformVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "第三方用户信息"
        ],
        "summary": "新增第三方用户信息",
        "operationId": "insertUsingPOST",
        "parameters": [
          {
            "in": "body",
            "name": "tbAccountPlatform",
            "description": "tbAccountPlatform",
            "required": true,
            "schema": {
              "originalRef": "TbAccountPlatformVO",
              "$ref": "#/definitions/TbAccountPlatformVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbAccountPlatformVO",
              "$ref": "#/definitions/TbAccountPlatformVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "第三方用户信息"
        ],
        "summary": "修改第三方用户信息",
        "operationId": "updateUsingPUT",
        "parameters": [
          {
            "in": "body",
            "name": "tbAccountPlatform",
            "description": "tbAccountPlatform",
            "required": true,
            "schema": {
              "originalRef": "TbAccountPlatformVO",
              "$ref": "#/definitions/TbAccountPlatformVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbAccountPlatformVO",
              "$ref": "#/definitions/TbAccountPlatformVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/account_platforms/{id}": {
      "get": {
        "tags": [
          "第三方用户信息"
        ],
        "summary": "通过ID查询单个第三方用户信息",
        "operationId": "findByIdUsingGET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbAccountPlatformVO",
              "$ref": "#/definitions/TbAccountPlatformVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "第三方用户信息"
        ],
        "summary": "通过ID删除单个第三方用户信息",
        "operationId": "deleteByIdUsingDELETE",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbAccountPlatformVO",
              "$ref": "#/definitions/TbAccountPlatformVO"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/admins": {
      "get": {
        "tags": [
          "管理员"
        ],
        "summary": "分页查询管理员",
        "operationId": "findByPageUsingGET_1",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "description": "页号",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页大小",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32",
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbAdminVO",
              "$ref": "#/definitions/TbAdminVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "管理员"
        ],
        "summary": "新增管理员",
        "operationId": "insertUsingPOST_1",
        "parameters": [
          {
            "in": "body",
            "name": "tbAdmin",
            "description": "tbAdmin",
            "required": true,
            "schema": {
              "originalRef": "TbAdminVO",
              "$ref": "#/definitions/TbAdminVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbAdminVO",
              "$ref": "#/definitions/TbAdminVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "管理员"
        ],
        "summary": "修改管理员",
        "operationId": "updateUsingPUT_1",
        "parameters": [
          {
            "in": "body",
            "name": "tbAdmin",
            "description": "tbAdmin",
            "required": true,
            "schema": {
              "originalRef": "TbAdminVO",
              "$ref": "#/definitions/TbAdminVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbAdminVO",
              "$ref": "#/definitions/TbAdminVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/admins/{id}": {
      "get": {
        "tags": [
          "管理员"
        ],
        "summary": "通过ID查询单个管理员",
        "operationId": "findByIdUsingGET_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbAdminVO",
              "$ref": "#/definitions/TbAdminVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "管理员"
        ],
        "summary": "通过ID删除单个管理员",
        "operationId": "deleteByIdUsingDELETE_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbAdminVO",
              "$ref": "#/definitions/TbAdminVO"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/commodity_attr_values": {
      "get": {
        "tags": [
          "销售属性值"
        ],
        "summary": "分页查询销售属性值",
        "operationId": "findByPageUsingGET_3",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "description": "页号",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页大小",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32",
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityAttrValueVO",
              "$ref": "#/definitions/TbCommodityAttrValueVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "销售属性值"
        ],
        "summary": "新增销售属性值",
        "operationId": "insertUsingPOST_3",
        "parameters": [
          {
            "in": "body",
            "name": "tbCommodityAttrValue",
            "description": "tbCommodityAttrValue",
            "required": true,
            "schema": {
              "originalRef": "TbCommodityAttrValueVO",
              "$ref": "#/definitions/TbCommodityAttrValueVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityAttrValueVO",
              "$ref": "#/definitions/TbCommodityAttrValueVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "销售属性值"
        ],
        "summary": "修改销售属性值",
        "operationId": "updateUsingPUT_3",
        "parameters": [
          {
            "in": "body",
            "name": "tbCommodityAttrValue",
            "description": "tbCommodityAttrValue",
            "required": true,
            "schema": {
              "originalRef": "TbCommodityAttrValueVO",
              "$ref": "#/definitions/TbCommodityAttrValueVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityAttrValueVO",
              "$ref": "#/definitions/TbCommodityAttrValueVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/commodity_attr_values/{id}": {
      "get": {
        "tags": [
          "销售属性值"
        ],
        "summary": "通过ID查询单个销售属性值",
        "operationId": "findByIdUsingGET_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityAttrValueVO",
              "$ref": "#/definitions/TbCommodityAttrValueVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "销售属性值"
        ],
        "summary": "通过ID删除单个销售属性值",
        "operationId": "deleteByIdUsingDELETE_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityAttrValueVO",
              "$ref": "#/definitions/TbCommodityAttrValueVO"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/commodity_attrs": {
      "get": {
        "tags": [
          "销售属性表 (产品参数)"
        ],
        "summary": "分页查询销售属性表 (产品参数)",
        "operationId": "findByPageUsingGET_2",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "description": "页号",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页大小",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32",
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityAttrVO",
              "$ref": "#/definitions/TbCommodityAttrVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "销售属性表 (产品参数)"
        ],
        "summary": "新增销售属性表 (产品参数)",
        "operationId": "insertUsingPOST_2",
        "parameters": [
          {
            "in": "body",
            "name": "tbCommodityAttr",
            "description": "tbCommodityAttr",
            "required": true,
            "schema": {
              "originalRef": "TbCommodityAttrVO",
              "$ref": "#/definitions/TbCommodityAttrVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityAttrVO",
              "$ref": "#/definitions/TbCommodityAttrVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "销售属性表 (产品参数)"
        ],
        "summary": "修改销售属性表 (产品参数)",
        "operationId": "updateUsingPUT_2",
        "parameters": [
          {
            "in": "body",
            "name": "tbCommodityAttr",
            "description": "tbCommodityAttr",
            "required": true,
            "schema": {
              "originalRef": "TbCommodityAttrVO",
              "$ref": "#/definitions/TbCommodityAttrVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityAttrVO",
              "$ref": "#/definitions/TbCommodityAttrVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/commodity_attrs/{id}": {
      "get": {
        "tags": [
          "销售属性表 (产品参数)"
        ],
        "summary": "通过ID查询单个销售属性表 (产品参数)",
        "operationId": "findByIdUsingGET_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityAttrVO",
              "$ref": "#/definitions/TbCommodityAttrVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "销售属性表 (产品参数)"
        ],
        "summary": "通过ID删除单个销售属性表 (产品参数)",
        "operationId": "deleteByIdUsingDELETE_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityAttrVO",
              "$ref": "#/definitions/TbCommodityAttrVO"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/commodity_brands": {
      "get": {
        "tags": [
          "品牌表"
        ],
        "summary": "分页查询品牌表",
        "operationId": "findByPageUsingGET_4",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "description": "页号",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页大小",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32",
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityBrandsVO",
              "$ref": "#/definitions/TbCommodityBrandsVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "品牌表"
        ],
        "summary": "新增品牌表",
        "operationId": "insertUsingPOST_4",
        "parameters": [
          {
            "in": "body",
            "name": "tbCommodityBrands",
            "description": "tbCommodityBrands",
            "required": true,
            "schema": {
              "originalRef": "TbCommodityBrandsVO",
              "$ref": "#/definitions/TbCommodityBrandsVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityBrandsVO",
              "$ref": "#/definitions/TbCommodityBrandsVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "品牌表"
        ],
        "summary": "修改品牌表",
        "operationId": "updateUsingPUT_4",
        "parameters": [
          {
            "in": "body",
            "name": "tbCommodityBrands",
            "description": "tbCommodityBrands",
            "required": true,
            "schema": {
              "originalRef": "TbCommodityBrandsVO",
              "$ref": "#/definitions/TbCommodityBrandsVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityBrandsVO",
              "$ref": "#/definitions/TbCommodityBrandsVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/commodity_brands/{id}": {
      "get": {
        "tags": [
          "品牌表"
        ],
        "summary": "通过ID查询单个品牌表",
        "operationId": "findByIdUsingGET_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityBrandsVO",
              "$ref": "#/definitions/TbCommodityBrandsVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "品牌表"
        ],
        "summary": "通过ID删除单个品牌表",
        "operationId": "deleteByIdUsingDELETE_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommodityBrandsVO",
              "$ref": "#/definitions/TbCommodityBrandsVO"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/commodity_sizes": {
      "get": {
        "tags": [
          "商品尺寸表"
        ],
        "summary": "分页查询sku表",
        "operationId": "findByPageUsingGET_6",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "description": "页号",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页大小",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32",
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommoditySizeVO",
              "$ref": "#/definitions/TbCommoditySizeVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "商品尺寸表"
        ],
        "summary": "新增sku表",
        "operationId": "insertUsingPOST_6",
        "parameters": [
          {
            "in": "body",
            "name": "tbCommoditySize",
            "description": "tbCommoditySize",
            "required": true,
            "schema": {
              "originalRef": "TbCommoditySizeVO",
              "$ref": "#/definitions/TbCommoditySizeVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommoditySizeVO",
              "$ref": "#/definitions/TbCommoditySizeVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "商品尺寸表"
        ],
        "summary": "修改sku表",
        "operationId": "updateUsingPUT_6",
        "parameters": [
          {
            "in": "body",
            "name": "tbCommoditySize",
            "description": "tbCommoditySize",
            "required": true,
            "schema": {
              "originalRef": "TbCommoditySizeVO",
              "$ref": "#/definitions/TbCommoditySizeVO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "TbCommoditySizeVO",
              "$ref": "#/definitions/TbCommoditySizeVO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "TbAccountPlatformVO": {
      "type": "object",
      "properties": {
        "avatar": {
          "type": "string",
          "description": "头像"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "id": {
          "type": "string",
          "description": "自增id"
        },
        "nickname": {
          "type": "string",
          "description": "昵称"
        },
        "platformId": {
          "type": "string",
          "description": "平台id"
        },
        "platformToken": {
          "type": "string",
          "description": "平台access_token"
        },
        "type": {
          "type": "integer",
          "format": "int32",
          "description": "平台类型 0:未知,1:facebook,2:google,3:wechat,4:qq,5:weibo,6:twitter"
        },
        "uid": {
          "type": "string",
          "description": "账号id"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        }
      },
      "title": "TbAccountPlatformVO",
      "description": "第三方用户信息模型"
    },
...
    "TbAdminVO": {
      "type": "object",
      "properties": {
        "avatarUrl": {
          "type": "string",
          "description": "管理员头像 (默认路径 “/avatar/boy.jpg”)"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "id": {
          "type": "string",
          "description": "管理员ID"
        },
        "mail": {
          "type": "string",
          "description": "管理员邮箱"
        },
        "name": {
          "type": "string",
          "description": "管理员用户名"
        },
        "password": {
          "type": "string",
          "description": "管理员密码"
        },
        "sex": {
          "type": "integer",
          "format": "int32",
          "description": "0表示男, 1表示女, 2表示未知"
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "description": "状态 1:enable, 0:disable, -1:deleted"
        },
        "superKey": {
          "type": "integer",
          "format": "int32",
          "description": "超级管理员标识 (0,普通管理员  1,超级管理员)"
        },
        "telephone": {
          "type": "string",
          "description": "管理员电话"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        }
      },
      "title": "TbAdminVO",
      "description": "管理员模型"
    },
...
  }
}